[SiFive][LV] Simplify the profitability runtime check
This change modifies LV's profitability runtime check to address one class of performance issues we have been experiencing related to this check. One class of these issues is a high overhead on the runtime check plus the scalar loop execution path (SCT-1912). The other class is profitable vector loops are bypassed due to this check's relatively high threshold, especially on our intelligence line. This change targets at the latter.

The change is checking against scalar iterations instead of vector iterations and have the threshold decoupled from LMUL. The new threshold is chosen based on @arcbbb's PM analysis, @nikolaypanchenko's Coremark analysis, @vishakha-agrawal's HPEC analysis, and my SPEC2006 int and GeekBench 5 data.

Other minor benefits of this change are (1) the threshold is no longer a runtime variable which is more friendly for optimizations; and (2) the check itself is slightly cheaper.