[SiFive][LAA] Update MaxSafeDepDistBytes when non-unit stride
MaxSafeDepDistBytes must be updated according to the stride
of the loop, otherwise it is possible that the value is
not conservative enough. For example:

```
for (int k = 0; k < len; k+=3) {
  a[k] = a[k + 4];
  a[k+2] = a[k+6];
}
```

has MaxSafeDepDistBytes=8. However, if we do not use the stride
in the calculation, we incorrectly compute MaxSafeDepDistBytes=24.

This change is related to 682cfc1, however, that change did not
go far enough and set MaxSafeDepDistBytes according to stride.