[SiFive][Merge] ValueTracking: Check instruction is in a parent in computeKnownFPClass
For some reason the inliner calls simplifyInstruction with disembodied
instructions. I consider this to be an API defect. Either the instruction
should always be inserted prior to simplification, or we at least
should pass in the new function for the context.

The patch is cherry-picked f42136d4d6e102659c6a4a870349aae6425a3ba9 from upstream.