[SiFive][RISCV] Use 4 byte alignment for arrays >= 16 bytes on RV32.
This allows us to hit the aligned access case in memcpy/strcmp/etc.
for such arrays.

We've used 8 byte alignment for array >= 8 on RV64 for a couple years now.
I went more conservative for RV32 because I only have dhrystone
performance data and >= 16 was the same perf as >= 8. My local
testing shows a 3% improvement on dhrystone for e76.