[SiFive][VPlan] CSADataUpdateRecipe uses correct data value
CSADataUpdateRecipe selected the new data based on which operand of
the select was a phi argument. It assumed that one argument was the
phi and one argument was the new data. But it did not consider if
the new data came from a different phi. This patch selects the correct
new data even if both arguments to the select are phis.