[SiFive][LV] Fix CSA when condition is negated
The CSA optimization wants to use a condition such that when it
is true, a new value is assigned. However, it is possible that a
true lane in WidenedCond corresponds to selection of the initial value
instead. In that case, we must use the negation of WidenedCond.

i.e. select cond new_val old_val versus select cond.not old_val new_val