[SiFive][LV] CSADataUpdate uses condition vector
Instead of using the NewMask, the CSADataUptate should use the
condition vector that guards assignment to the new value. The reason
for this is because if the condition vector is all false but NewMask
has an active lane in 0..RVL, then the data will be updated from the
current iteration when the data should remain from the previous
iteration.