[SiFive][RISCV] Properly handle partial writes in isConvertibleToVMV_V_V.
We were only checking for the previous writer to write exactly
the register or multiple registers. We ignored partial writes and
continued scanning for other writers. We need to abort for partial
writes.

There's a later check inside the if body to abort if the registers
don't match exactly. So we just need to check for overlap so we
enter the if body.

I don't have a good reduced test case for this right now. I'll
try to make an MIR case before I upstream this.

Fixes SCT-2687.