[SiFive][ASAN] Introduce TTI::getMemoryRefInfo and RVV support. (#1234)
The patch is a proof of concept to create a hook to make AddressSanitizer to
handle target intrinsics for memory references.
The function getMemoryRefInfo(Intrinsic) will return a MemoryRefInfo object of
the input intrinsic. MemoryRefInfo is the new name of InterestingMemoryOperand.
InterestingMemoryOperand was a data structure to track information of memory
references and used by AddressSanitizer and HWAddressSanitizer.

A part of the patch is extracted form D149021, but we don't rename
InterestingMemoryOperand to MemoryRefInfo for easier maintaince.

The patch consider one segment load/store as multiple InterestingMemoryOperands.