[RISCV] More accurately model the bypass between loads and load/store for SiFive7.
According to the bullet MAS

"Loads produce their result in the M2 stage. There is no load-use delay for most integer instructions. However, effective addresses for memory accesses are always computed in the AG stage. Hence, loads, stores, and indirect jumps require their address operands to be ready when the instruction enters AG. If an address-generation operation depends upon a load from memory, then the load-use delay is usually two cycles. However, if the dependent address-generation operation is a load or store instruction with an offset of exactly zero, the load-use delay is only one cycle."

So I think we should have 3 cycle latency unless the load/store offset is 0,
then we can have 2 cycle latency.