[SiFive][SCT-2513][LV] Fix incorrect use of vscale in loop
I learned that RVL is not always VLMAX in loops when VLMAX
< AVL < 2*VLMAX. According to the V Spec, a range like ceil(AVL/2) <=
RVL <= VLMAX is allowed. So we need to be cautious when using vscale in
the loop.

To address this issue, this patch moves the step calculation from the
preheader to the loop body since it is now loop variant, and utilizes
RVL instead of vscale for scalable VF.