[SiFive][RISCV] Stop performFP_TO_INTCombine from folding with ISD::FRINT.
Cherry-picked from D157583 to fix SCT-2867.

Upstream commit message:

FRINT was added to matchRoundingOp after this function was written.
So FRINT was not tested originally.

For vectors, folding this causes us to create a CSR swap that tries
to write 7 to FRM. This is an illegal value and will cause the CSR
write to fail.

While this might be a legal fold we could do, I'm disabling it for
now so we can backport to LLVM 17 with the least risk.