[SiFive][LV] Fix VPCSADataUpdateRecipe merge operands
The VPCSADataUpdateRecipe uses the operands of the select that it was
built from. Prior to this patch, NewData was built by selecting
UndistData when Cond is true and OldData when Cond is false. However, it
should be built to select UndistData when the select would have selected
new data and select OldData when the select would have selected the old
data.