[SiFive][LV] Adjust cost of `VPReplicateRecipe` (#1353)
The `VPReplicateRecipe` represent scalar instructions: if value is
uniform, only one instruction will be generated; if value is non-uniform,
scalarization is done. The latter is not yet supported by VLA
vectorization.

The changeset adjusts the cost of `VPReplicateRecipe` to be `1` if value
is uniform and use `ElementCount` for now otherwise, which should not be
triggered with current implementation.