[SiFive][InstCombine] Try to constant fold terminator of vsetvli loops emittered by our vectorizer.
If the scalar trip count becomes constant after vectorization, we might
not realize the loop only needs 1 vector iteration.

This patch pattern matches the idiom our vectorizer uses and
constant folds it if we can prove the first AVL is less than or
equal to VLMAX of the vsetvli.

This strips some control flow from loops in 505.mcf_r.

Part of SCT-2969.