[SiFive][RISCV] Optimize (abs (sub (zext X), (zext Y))) -> (zext (sub (zext (max X, Y), (min X, Y)))).
Where abs is an i32 vector and X and Y are i8 vectors. If the use
is an add or a reduction add.

The max and min done first increase instruction level parallelism.
The moving of the zexts allows us to make use of vwsubu and vwadd
to produce 2x DLEN bits per cycle.