[SiFive][SLP]Fix SCT-3023: do not crash when trying to extract (#1461)
non-vectorized scalars for strided loads pointers.

If the load node is determined to emit strided load, not a masked
gather, there is no need to vectorize its pointer operand and their
operands too. It may lead to the fact, that some nodes gets not
vectorized and when trying to extract external scalars, the compiler
will crash.