[SiFive][Merge] Cherry-picked PR#67488
Fix a crash from trying to truncate an FP type in lowerBuildVectorOfConstants.
ComputeNumSignBits can return an answer for FP constants based on bitcasting them to int.
Check for an integer type so we don't create an illegal truncate.
We could support this case with bitcasts, but I leave that to a separate patch.