[SiFive] Simplify combineOrZextShlAnyext. NFC
Check for fixed length vector at the beginning instead of relying
on the BUILD_VECTOR opcode to block scalable vectors.

Handle the commuted operand check for the OR with a std::swap to
remove the for loop.

Simplify how we detect the splat shift amount.

Keep an unsigned for the EltSize that we can use to detect the other
element sizes.

Check the correct Hi VT by just comparing to the Lo VT we already have.

Pre size the SmallVector to avoid push_backs that might need to resize.
This should reduce the binary size of the compiler too since we don't
need to check if the vector needs to grow.

Rename variables to remove the word Result and shorten "EVT" to "VT"
which is the usual SelectionDAG naming convention.