[SiFive][TypePromotion] Add a small hack to type promotion to prevent adding zext for some signed values.
We have enabled the ARM TypePromotion pass in our downstream. This pass
looks for compares on small illegal types like i8 and i16. It tries to
rewrite graph to widen these illegal types to XLen and place the zext
at a more optimal place.

The pass considers at eq/ne comparison as being unsigned and tries
to insert zext earlier. If the value happens to also be used by a
signed compare this will result in a trunc+sext before the signed
icmp. This results in both sext and zext existing in the graph in
some cases. If the producing instructon is a load, its arbitrary whether
we get a sextload or zextload. The other users will get a zext.h or sext.h.

SelectionDAG also has some heuristics for how it extends small live outs of
basic blocks based on whether the value is ever used by a signed compare.

This patch adds a simple bail out to the TypePromotion pass for cases
where we can easily see the value should be sign extended. We'll fall
back to the SelectionDAG heuristics.

Fixes SCT-3034.