[SiFive] Rework TypePromotion hack to avoid regressing AArch64 test.
Prior to my change the AArch64 test called TryPromote based on the
icmp of a load and then blocked due to some other heuristics. With my check
it no longer promotes on the load operand, but instead promotes
for the phi operand. The other heuristics are disabled if the source
node is a PHI. That heuristic might be fragile because I think the
PHI check was for promotions originating from a loop phi, not a PHI
that is an operand of an icmp.

This patch moves my bail out to the end of TryPromotion after we've
collected all the candidates.