[SiFive][JumpThreading] Improve jumpthreading if predecessor has an identical compare after phi translation.
If we're branching on a compare with a PHINode in a block, and
and one of its predecessors has a similar compare and branch. We know
the compare is true in this block if this block is the true successor
of the predecessor.

We can use this to make the predecessor branch skip branching to
this block and go straight to this block's successor.

Fixes SCT-3112.

TODO: Extract test case.