[SiFive][Analysis][LiveValues] Value Pressure query support for LICM hoisting
This change contains support to calculate IR Value Pressure for a number of
RISC-V register classes in llvm IR.  This support utilizes data flow
sets and live interval data collected by the loop pass manager analysis
of the current function. As this is a temporal analysis which is not done
for each loop optimization pass, the initial data can become stale on
subsequent optimization passes, ergo we need checks for both data that
is no longer resident and for resident data that is not present in that
analysis.  Both need to screened while calculating Value Pressure
for a given set of blocks collected by the query function. Initially
the early loop optimization passes invoked by the loop pass manager are
very close if not precisely accurate wrt data flow and live interval data.
The first support context is for LICM hoisting.  These queries aid in
guarding optimization in cases where we have estimated register pressure
to be sufficiently high to induce spill code generation in the code generator.
In cases where Value Pressure is over the machine limit for a given Register
Class, we screen candidates that only improve Value Pressure if code motioned.
Additional studies will need to be made to see if sycning in LICM will
also yield any benefits to guard against optimization.  Other candidates
may also include GVN and its variants.