[SiFive] Early expand vp.reduce.fadd with non-zero VL to improve generated code
The RVV reduction instructions don't update their destination
register if the VL is 0. The VP reduction intrinsics are defined
to copy their scalar input to the result when the VL is 0.

To satisfy the VP behavior with instructions, SelectionDAG puts
the scalar value in the passthru operand and the scalar operand.
This way if VL is 0, the scalar value is already in the destination
register through the passthru operand.

This does mean if there are 2 reductions that use same scalar, we
will end up with a copy of the scalar so that it can be in a
different destination register for the other reduction.

The other issue is that VP intrinsic needs to create a vfmv.s.f
instruction to put the scalar in a register. vfmv.s.f also doesn't
update the destination register when VL is 0. To ensure the vector
always gets updated, SelectionDAG will force the VL to 1. This
results in an extra VL toggle.

This patch introduces a IR expansion that uses isKnownNonZero to
create an more optimal sequence. We can set the passthru operand
of the vfredusum to poison and use the non-zero VL for the vfmv.s.f.

We can extend this to other reductions in future patches.