[SiFive][InstCombine] Convert vp.merge to vp.select if the tail isn't used.
I'm only doing this if the vp.merge is the true operand of another
vp.merge or the true/false operands of a vp.select. We can extend
this to other vp intrinsics in the future.

My initial goal was to clean up chains of vp.merges. Though maybe

The vp.merge+vp.merge transform is now updated to look for a vp.select
operand. The new transform ensures we would never see 2 vp.merges.
This change is also needed after #1620.