[SiFive][InstCombine] Swap operands of vp.select if mask is a vp.not.
We can bypass the vp.not and swap the operands. This may put the
false value in the right place for the vp.select to be combined
with a later vp.merge. A pattern I have seen several times.

I will probably add another combine to commute a vp.select and add
a vp.not if it will allow a vp.merge and vp.select to to be combined.