[SiFive][MISched] Handle empty sized resource usage.
TargetSchedule.td explicitly allows the usage of a ProcResource for zero
cycles, in order to represent that the ProcResource must be available
but is not consumed by the instruction. On the other hand,
ResourceSegments explicitly does not allow for a zero sized interval. In
order to remedy this, this patch handles the special case of when there
is a zero sized usage of a resource without creating a ResourceSegment
for it.

This is the downstream version of
https://github.com/llvm/llvm-project/pull/75951.