[SiFive][LoopProfiler] Optimize string creation. NFC
The original code called operator+ on multiple std::strings which may
require multiple heap allocations. Instead, build the string first using
Twine, then flatten the Twine to a std::string.

Constructing a Twine doesn't allocate any memory on the heap. It will
build a tree of pointers to individual std::strings or StringRefs that
exist elsewhere. A Twine can also store an integer. This avoids copying
any of the strings while building up the Twine. Then we flatten the
Twine to a std::string which will accumulate the sizes of all pieces
before creating the std::string so it knows exactly how much to
allocate. Thus reducing the number of heap allocations needed.