[SiFive][RISCV] Optimize storing vector.interleaved2 of constant splats (#1639)
Currently InterleavedAccessPass will turn this pattern into segment
store, which is quite expensive. This patch turns the original
vector.interleaved2 into a larger splat and bitcast it before feeding
back to the original vp.store.
This will result a unit-stride store, which is much more efficient in
this case.

SCT-3273.