[SiFive][RISCV] Add sifive-x392-ea-dualvalu, and sifive-x392-ea-singlevalu
There are two versions of the sifive-x390: a single valu version and a
dual valu version. However, the product name does not differentiate
between the two. For this reason, we give the mcpu names
sifive-x392-ea-singlevalu and sifive-x392-ea-dualvalu. The `ea` stands
for the early access release.

This patch gives all mcpus the SiFive7 scheduler model for now, since
SiFive7 is based off of Bullet and the X390 is a Bullet processor.
Future patches will make modifications so that the scheduler model of
the x390 CPUs is correctly tuned (it is likley that we will create a new
model for the x390 processors).