[SiFive][RISCV][NFC] Begin to reuse classes that can be shared by bullet based processors
We have two Bullet based SchedModels at the moment, and plan to add more
with the X390 family. Some of the classes in the existing scheduler
models can be factored out as commonality due to processors being part
of the Bullet micro-architecture.

I plan to factor out even more as we go forward, but figured this was
the most basic start.