[SiFive][RISCV] Add SiFive7SSingleVALUModel and SiFive7SDualVALUModel
There are x392 configurations that have two VALUs. The two VALUs are not
symmetric. I need to double check that I partitioned the instructions
correctly with respect to what VALU they can use before this patch is
ready to merge. The x392 has the same latencies as the SiFive7O for now.
There may be improvements to come, but we will model once those
optimizations oficially make it into the processor.