[SiFive][RISCV] Factor out WriteRes and ReadAdvance common to Bullet. NFC.
Each CPU entry specifies a scheduler model to use. Prior to this patch,
the code reuse between similiarly behaved CPUs was quite poor and this
has lead to duplicated work and even bugs.

In anticipation of adding the x390 processor family, this patch factors
out WriteRes and ReadAdvance records that are common to all the Bullet
processors. The WriteRes and ReadAdvance records that are different
between different CPUs should be defined in the derived class. In the
future, if a record should have different behavior, it should be moved
out of the Bullet base class and into the SchedModels.

The code containing variants could be refactored in future patches, but
I had some trouble figuring it out so I left those alone in this patch.

We should consider making a Mallard version of this patch.

I think that I'd probably like to split the x280n and x280o into their
own models in the future patch, since the files would have little
duplicate logic, and I prefer to not use variants if we can avoid it.