[SiFive][RISCV] Make XSfvfhbfmin alias Zvfbfmin.
Xsfvfhbfmin uses the same encodings as Zvfbfmin, but with different
mnemonics.

As a step towards migrating away from XSfvfhbfmin, this patch makes
Xsfvfhbfmin imply Zvfbfmin and makes the instructions aliases in the
assembler. I've also kept the aliases for the printer.

I've also restored upstream code that makes XSfvfwmaccqqq imply
Zvfbfmin. This allows us to remove customization in SemaChecking.cpp.