[SiFive][RISCV] Add new MI pass to form SHxADD instructions across basic blocks. (#1711)
LSR term folding is creating cases where we have a shl in the preheader
of an outer loop used by a GEP the preheader of an inner loop.

In the cases I've seen, the shl is using the value holding the trip
count of the inner loop which is invariant to the outer loop. This
results in both the trip count and the shifted trip count being live
into the outer loop increasing register pressure.

This patch adds a new pass to use the Zba SHxADD(_UW) instructions to
sink the SLLI into the outer loop by combining it with an ADD
instruction created by the GEP.

Test cases are a fairly basic shift with an add in a loop rather than
the exact case I saw in benchmarks. At the level we're doing this
transform the difference shouldn't matter.

This recovers the dynamic instruction count on 483.sphinx3 that
regressed in week 127.