[SiFive][RISCV] Prefer vlse with x0 stride to vfmv.v.f on x280 (#1704)
# Overview

Other instructions may be keeping vector arithmetic sequencer busy, and
a vmv.v.x may wait in the VCQ until the arithmetic sequencer is free.
However a vlse instruction loading from a constant pool could have been
emitted instead, which uses the load sequencer. If the load sequencer is
free, emitting a vlse can lead to better utilization of the vector unit.

# Background
The memory access for a vlse is started in the scalar pipe before the
load enters the VCQ. For strided loads, all the address calculations
happen in the scalar pipeline – the 1 cycle per element occurs in the
scalar pipeline. In the vector pipeline, the write from the load buffer
to the vector register occurs. This means the memory latency can overlap
with other operations before the load reaches the head of the VCQ. On
their own, with an empty pipeline, vmv.v.x and vlse both take
NumDLENParts cycles.

# Solution

A vlse instruction is always emitted in favor of a vfmv.v.f This
solution is proposed due to the empirical results collected in spec2006:

|benchmark | base_score | opt_score | opt_speedup|
|---|---|---|---|
|433.milc | 57834.04 | 57608.46 | 0.3900471072053824| |453.povray |
43501.55 | 43507.5 | -0.013677673554157721|
|444.namd | 72316.56 | 72578.87 | -0.3627246650006466| 
|447.dealII | 53952.1 | 53391.02 | 1.0399595196479927|
|470.lbm | 34036.78 | 33744.58 | 0.8584830880006833|
|450.soplex | 77163.7 | 77065.11 | 0.12776733101185211|

# Implementation

## Summary

Entering into SelectionDAG there are multiple ways a splat vector can be
represented:
  1. ISD::SPLAT_VECTOR
  2. ISD::BUILD_VECTOR where all elements are the same
  3. RISCVISD::VFMV_V_F_VL
  4. Intrinsic::riscv_vlse using x0 stride

In the SelectionDAG pipeline, the combine pass is run multiple times
before instruction selection runs. As a result, it is likely that the
code contains primarily SPLAT_VECTOR, and sometimes BUILD_VECTOR, or the
RISC-V specific VFMV_V_F_VL or Intrinsic::riscv_vlse. We should rarely
see the BUILD_VECTOR case since it would have been more appropriate to
build a SPLAT_VECTOR in the first place. The only way to see RISC-V
specific nodes would be due to intrinsics or if a prior combine pass
created a RISC-V specific node.

After the combine passes run, DoInstructionSelection calls
PreprocessISelDAG. Here, ISD::SPLAT_VECTOR is converted to VFMV_V_F_VL.

In the Select function, we attempt to convert VFMV_V_F_VL into
riscv_vlse instructions with x0 stride. The conversion only happens if
the scalar argument is a load, some cost calucations, and some other
requirements that are not relevant to our context. The key point is that
we'd like to reuse this logic to solve our problem.

## Changes Made

The main goal is to convert SPLAT_VECTOR whose scalar operand is a
ConstantFPSDNode into a SPLAT_VECTOR with a scalar operand that is a
load from the constant pool. That way, PreprocessISelDAG will convert
the SPLAT_VECTOR to a VFMV_V_F_VL whose scalar operand is a load from
the constant pool. That will cause the existing code in Select to
convert VFMV_V_F_VL into iscv_vlse instructions with x0 stride.

One alternative option would be to convert the SPLAT_VECTOR to a
VFMV_V_F_VL during the combine phase, but then the combine phase would
miss out on optimization opportunities that expect a SPLAT_VECTOR
instead of a VFMV_V_F_VL.

We can also convert VFMV_V_F_VL whose scalar operand is a
ConstantFPSDNode into a VFMV_V_F_VL whose scalar operand is a load from
the constant pool. That will allow us to reuse the code in Select to
convert RISCVISD::VFMV_V_F_VL into riscv_vlse instructions with x0
stride.

The Intrinsic::riscv_vlse case is already in the desired form.

This patch does not address the BUILD_VECTOR case.