[SiFive][JumpThreading] Handle unreachable dest in isSwitchLookupTable.
Upstream recently made a change to CVP that caused the default case
in the switch I added this code for to become unreachable.

This caused jump threading to start threading it again, breaking
the pattern CFG is looking for. Only the last run of SimplifyCFG in
the pipeline is able to create lookup tables.

This regressed dhrystone because a function in libmetal now generates
a large lookup table that blew up the size of the data section. This
prevented the global pointer from being near the global variables in
actual dhrystone program. The linker script is kind of bad and should
probably have more control of where the global pointer is, but that
requires fixes on multiple branches of federation. So I'm fixing the
compiler again instead.