[SiFive][LV][SCT-3158] Allow ptr induction and low trip count vectorization (#1542)
The changeset enables pointer IV which is necessary for `satd_WxH`
vectorization in `525.x264` benchmark.
Simple enabling degraded the performance of the benchmark, due to
vectorization of loops with non-constant strides, where SLP is doing
much better job as it can use unit-strided loads/stores.
To mitigate it, the changeset also adds special heuristic which tries to
predict if SLP will vectorize the loop body. The heuristic uses
`InterleavedAccessInfo` to make a decision and to make it work for all
important cases in `525.x264` it's also required to:
* increases `max-interleave-group-factor` to 16 to group accesses in
`sad_16x16`. This won't impact performance, but will impact compile
time.
* rebuild groups within the heuristic as if no versioning for strides is
done.