[SiFive][RISCV] Try to make mask instructions avaliable for scheduling sooner.
On SiFive7, mask instructions scheduled too close to their uses expose
the latency of the arithmetic pipeline to the subsequent dependent
instruction and causes the arithmetic sequencer to stall for a number of
cycles.

The goal of this patch is to make the scheduler choose available mask
instructions over other instructions that are sooner in the original instruction
ordering.

This is accomplished by using a ScheduleDAGMutation where a weak edge is
added from the mask instructions to the instructions before it in the
scheduling region. The weak edge is an arbitrary weak DAG edge which
signifies that there is no real dependence. In other workds, a weak edge acts
acts as a hint to the scheduler to prefer the source node over the sink
node.

If there is a real dependence between the sink node, then the weak edge
is ignored. This allows us to add the weak edge to all prior
instructions without having to calculate whether the weak edge should
exist or not.