[SiFive][RISCV] Add RISCVMaskInstrHazardRecognizer
On SiFive7, mask instructions operate at EMUL=1, so when a dependent
instruction is too close to the mask instruction, the latency of the
arithmetic pipeline is exposed.

Add a HazardRecognizer that models that it is a hazard for a mask
instruction to be scheduled next to any of its uses.

The purpose of this patch is to allow instructions that are pending to have time
to enter the Available queue to be scheduled, in hopes that the
scheduler will have an instruction to put in between the mask
instruction and its consumer.

We hardcode an 8 cycle bailout to the hazard which is intended to model
that if there is no instruction avaliable after 8 cycles, it is okay to
schedule the mask instruction and incur the exposed latency.