[RISCV] Optimize (srl (bswap X), 32+C) to (SRLI (SRLI (BSWAP X), 32), C).
If (SRLI (REV8 X), 32) will be the only use of BSWAP and has at least 4
uses.

This is a hack to reduce register pressure in a workload with 16 such
patterns. This reduces the number of live registers, by making sure we
don't need to keep the REV8 register live past the SRLI 32.