[SiFive][LV][SCT-3526] Correctly extract invariant stride for IG (#1773)
Pointer increment can be done in a different type than access, which
could also mean its stride can be already in bytes. The changeset makes
analysis of IAG to use `StrideAccessInfo` that is used to generate
strided intrinsics. Unlike strided intrinsics, for IAG the stride is
converted to element strides.

The changeset also fixes potential bug with `StrideAccessInfo` when
runtime stride is replaced by `1`.