[SiFive][RISCV] Fix Bullet scheduler for complex mask instructions
VIOTA and VID have the standard occupancy of 2*LMUL and 4 cycle latency.
The other complex mask instructions have latency and occupancy of 11 cycles.
This information was taken from Eric Love's spreadsheet. Jaehoon is
going to update the spreadsheet for VIOTA and VID at the time this
commit was written since it incorrectly reported they had the 11 cycle
latency and occupancy