[SiFive][RISCV] Increase large array threshold for riscv64 to match riscv32.
This threshold is intended to help align arrays so they can use aligned
memcpy/memset/strcpy/etc or ld/sd. The extra padding can have a negative
effect on the layout of global variables and limit the use of GP
relaxation.

This increases the threshold so we align 16 byte and large arrays
instead of 8 byte. I checked the original motivating case and its a 44
byte array that is partially vectorized by SLP today.