[SiFive][LV] Don't pass all true mask to the first operand of vp_popcount
Passing an all true mask causes the backend to emit a vmset to
create an all true mask. Moving the mask from the mask operand to
the first operand allows the backend to use the unmasked form of
vcpop.m saving a vmset.

We could optimize this in the backend, but it was easier to fix
at the source.