[SiFive][InstCombine] Optimizing reduction tree (#1855)
SLP + Unroll can result to a sequence of vector reduces that are
accumulated into same scalar value in some non-trivial way. For example
```
%start0   %v0
   \     /
  vecreduce       %0  %start1 %v1
        \       /       |      /
           add      vecreduce
            |     /
           add
```
the changeset implements optimization in InstCombine to walk over
use-def and def-use chains to find vector reduces in computation tree
that can be optimized into vector accumulator and single vecreduce. That
is, for the example above it optimizes it into:
```
                            %v0  %v1
                              \  /
    %start0      %0  %start1  add
        \       /       |      /
          add      vecreduce
           |     /
          add
```