[SiFive][LV] Fixing the dominator tree for re-vectorization without SCEV checks (#1872)
There are two versions of vectorized loops if enabling symbolic stride
memory access versioning: the vectorized loop with SCEV checks and the
re-vectorized loop without SCEV checks. However, the requirements for
whether an epilogue is needed may differ between these two versions of
the vectorized loop. This patch fixes the updating for the dominator
tree when the vectorized loop with SCEV checks requires an epilogue but
the re-vectorized loop don't.

SCT-3658