[SiFive][RISCV] Remove IsEABI from RISCVZC::getStackAdjBase. (#89177)
SIFIVE cherry-pick from cd0c94bb8558b6e26313694da18006a840521903

The usage of IsEABI was only valid for RV32E. For RV64E, the stack
adjust
base needs to be 32 when ra,s0-s1 are being saved. Since it takes more
than 16 bytes to save 3 64-bit registers.

The spec lists the rv32e behavior explicitly, but not rv64e. My
assumption is that the only thing that changes with rv64e is which
registers can be used in the register list, but not how the register
list affects the stack_adj_base.