[SiFive][RISCV] Add extension information to RISCVFeatures.td. NFC (#89326)
SIFIVE cherry-pick from a6c028299cbce9885dd3c3e989b2cb3273ba6e05

This adds a new RISCVExtension class that inherits from
SubtargetFeature. This contains the major/minor version and whether the
extension is experimental. The plan is to use this to generate the
tables for RISCVISAInfo.cpp.

The version numbers might not be accurate yet. If there are errors they
will be fixed before they are used for anything. It will be easier to
verify once the new tablegen backend is written to generate the
RISCVISAInfo.cpp table.